+++ imports/startup/both/routes.js
@@ -10,6 +10,7 @@ import { urlDoctor, toTitleCase } from '/lib/utils';
 import { Contracts } from '/imports/api/contracts/Contracts';
 import { stripHTMLfromText } from '/imports/ui/modules/utils';
 import { displayNotice } from '/imports/ui/modules/notice';
+import { shortenCryptoName } from '/imports/startup/both/modules/metamask';
 import { tokenWeb } from '/lib/token';
 import { sync } from '/imports/ui/templates/layout/sync';
 import { request } from 'http';
@@ -152,9 +153,20 @@ Router.route('/address/:username', {
   onBeforeAction() {
     Session.set('sidebarMenuSelectedId', 999);
     _reset();
+
+    Session.set('search', {
+      input: '',
+      query: [
+        {
+          id: this.params.username,
+          text: shortenCryptoName(this.params.username),
+        },
+      ],
+    });
+
     this.next();
   },
-  data() {
+  data() {    
     return {
       options: { view: 'peer', sort: { timestamp: -1 }, limit: gui.ITEMS_PER_PAGE, skip: 0, username: this.params.username },
     };
